NEUBURG-2: Move database/* to models/* and cleanup
