+++ src/services/initializeServices.js
@@ -1,8 +1,6 @@
 import OfferService from './OfferService'
 
 export default () => {
-  const offersFileName = './offers.json'
-  const offerService = new OfferService(offersFileName)
-
+  const offerService = new OfferService()
   return {offerService}
 }
